+++ react/B2BAdminLayout.tsx
@@ -1,12 +1,16 @@
 import type { FunctionComponent } from 'react'
 import React from 'react'
-import { createSystem } from '@vtex/admin-ui'
+import { createSystem, ToastProvider } from '@vtex/admin-ui'
 import 'vtex.country-codes/locales'
 
 const B2BAdminLayout: FunctionComponent = ({ children }) => {
-  const [ThemeProvider] = createSystem({ key: 'btob-organizations' })
+  const [SystemProvider] = createSystem({ key: 'btob-organizations' })
 
-  return <ThemeProvider>{children}</ThemeProvider>
+  return (
+    <SystemProvider>
+      <ToastProvider>{children}</ToastProvider>
+    </SystemProvider>
+  )
 }
 
 export default B2BAdminLayout